Get SNOMED codes with a specific set of one or more attributes

Optionally filtered for specific relationship types. See examples (credit to snomedizer).

Usage

HAS_ATTRIBUTES(
  attribute_codes,
  relationship_type = NULL,
  standardise_output = TRUE,
  all_lkps_maps = NULL,
  preferred_description_only = TRUE,
  col_filters = getOption("codemapper.col_filters")
)

Arguments

attribute_codes

Character vector of SNOMED codes.

relationship_type

Character vector of SNOMED codes.

standardise_output

bool. If TRUE (default), outputs a data frame with columns named 'code', 'description' and 'code_type'. Otherwise returns a data frame with all columns from the relevant look up table.

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a Duckdb database containing these tables. If NULL, will attempt to connect to a Duckdb database named 'all_lkps_maps.db' in the current working directory, or to a a Duckdb database specified by an environmental variable named 'ALL_LKPS_MAPS_DB' (see here for how to set environment variables using a .Renviron file). The latter method will be used in preference.

preferred_description_only

If TRUE (default), return only preferred descriptions for clinical codes with synonyms. Will only apply if standardise_output is also TRUE.

col_filters

A named list where each name in the list refers to the name of a lookup or mapping table. Each item is also a named list, where the names refer to column names in the corresponding table, and the items are vectors of values to filter for. For example, list(my_lookup_table = list(colA = c("A", "B")) will result in my_lookup_table being filtered for rows where colA is either 'A' or 'B'. Uses default_col_filters() by default. Set to NULL to remove all filters.

Value

A dataframe

See also

Examples

if (FALSE) { # \dontrun{
# Conditions associated with multiple sclerosis
HAS_ATTRIBUTES("24700007", relationship_type = "42752001")

# Medicines with active ingredient timolol maleate
HAS_ATTRIBUTES("75359005", relationship_type = "10362801000001104")

# Medicines with active ingredient beta blocker
HAS_ATTRIBUTES(CHILDREN("373254001", code_type = "sct"), relationship_type = "10362801000001104")

# Conditions that are caused by bacteria belonging to Enterobacteriaceae
HAS_ATTRIBUTES(CHILDREN("106544002", code_type = "sct"), relationship_type = "246075003")

# Infectious conditions that are caused by bacteria belonging to Enterobacteriaceae
HAS_ATTRIBUTES(CHILDREN("106544002", code_type = "sct"), relationship_type = "246075003") %AND%
  CHILDREN("40733004", code_type = "sct")
} # }