Shiny app for building clinical codelists.

Launches a Shiny app. Note, requires all_lkps_maps database file.

Usage

RunCodelistBuilder(
  all_lkps_maps = NULL,
  options = list(launch.browser = TRUE),
  ...
)

Arguments

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a Duckdb database containing these tables. If NULL, will attempt to connect to a Duckdb database named 'all_lkps_maps.db' in the current working directory, or to a a Duckdb database specified by an environmental variable named 'ALL_LKPS_MAPS_DB' (see here for how to set environment variables using a .Renviron file). The latter method will be used in preference.

options

Named options that should be passed to the runApp call (these can be any of the following: "port", "launch.browser", "host", "quiet", "display.mode" and "test.mode"). You can also specify width and height parameters which provide a hint to the embedding environment about the ideal height/width for the app.

...

Additional args passed on to shinyApp

Value

NULL

Examples

if (FALSE) { # \dontrun{
# build dummy all_lkps_maps database
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()
all_lkps_maps_db <- all_lkps_maps_to_db(
  all_lkps_maps = all_lkps_maps_dummy,
  db_path = tempfile(fileext = ".db")
)

# launch app
RunCodelistBuilder(all_lkps_maps = all_lkps_maps_db)
} # }