Get attributes for related SNOMED codes

Low level function for querying related SNOMED codes, returning attributes types.

Usage

get_attributes_sct_old(
  codes,
  relationship = NULL,
  relationship_direction = "child",
  recursive = TRUE,
  active_only = FALSE,
  col_filters = getOption("codemapper.col_filters"),
  all_lkps_maps = NULL
)

Arguments

codes

Character vector of SNOMED codes

relationship

Character vector of SNOMED codes defining the types of relationship. If NULL (default), returns all types of relationship.

relationship_direction

Either 'child' (default) or 'parent'.

recursive

If TRUE (default), will recursively search for related codes e.g. find all descendants code instead of just the immediate child codes.

active_only

If FALSE (default), return all relationships, even if currently inactive.

col_filters

A named list where each name in the list refers to the name of a lookup or mapping table. Each item is also a named list, where the names refer to column names in the corresponding table, and the items are vectors of values to filter for. For example, list(my_lookup_table = list(colA = c("A", "B")) will result in my_lookup_table being filtered for rows where colA is either 'A' or 'B'. Uses default_col_filters() by default. Set to NULL to remove all filters.

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a Duckdb database containing these tables. If NULL, will attempt to connect to a Duckdb database named 'all_lkps_maps.db' in the current working directory, or to a a Duckdb database specified by an environmental variable named 'ALL_LKPS_MAPS_DB' (see here for how to set environment variables using a .Renviron file). The latter method will be used in preference.

Value

A data frame

See also

Examples

if (FALSE) { # \dontrun{
# get children codes for "269823000" ("Hemoglobin A1C - ...")
get_attributes_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "child"
  )

# get parent codes
get_attributes_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "parent"
  )

# get all codes for which "386868003" ("Bisoprolol (substance)") is an attribute.
# Note, includes "293967003" ("Allergy to bisoprolol (finding)")
get_attributes_sct(
  codes = "386868003",
  relationship = NULL,
  recursive = FALSE,
  relationship_direction = "child"
  )

# more results (i.e. bisoprolol-containing medications) are returned with `recursive = TRUE`
get_attributes_sct(
  codes = "386868003",
  relationship = NULL,
  recursive = TRUE,
  relationship_direction = "child"
  )
} # }