Reformat ICD-10 codes

The lookup sheet in UKB resource 592 for ICD-10 ("icd10_lkp") has a column called ALT_CODE, which is an alternative format for ICD-10 codes. This is the format used in the mapping sheets for this resource, as well as in Field ID 41270. Note however that in contrast to the ICD10 lookup and mapping sheets, undivided 3 character codes do not have an 'X' appended in the UK Biobank dataset. This function converts from one format to the other, with the option to strip the final 'X' from undivided 3 character codes if converting to ALT_CODE format.

reformat_icd10_codes(
  icd10_codes,
  all_lkps_maps = NULL,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE",
  unrecognised_codes = "error",
  strip_x = FALSE
)

Arguments

icd10_codes

character vector of ICD-10 codes

all_lkps_maps

Either a named list of lookup and mapping tables (either data frames or tbl_dbi objects), or the path to a SQLite database containing these tables (see also build_all_lkps_maps() and all_lkps_maps_to_db()). By default, will look for a SQLite datbase called all_lkps_maps.db in the current working directory.

input_icd10_format

character. Must be either "ICD10_CODE" or "ALT_CODE".

output_icd10_format

character. Must be either "ICD10_CODE" or "ALT_CODE".

unrecognised_codes

Either 'error' (default) or 'warning'. If any input codes are unrecognised, then either an error or warning will be raised.

strip_x

If TRUE and converting to ALT_CODE format, 'X' is removed from the end of undivided 3 character codes (default is FALSE).

Value

character vector of ICD-10 codes, reformatted as specified by output_icd10_format.

See also